Revert "Super Bowl Update"
