+++ README.md
@@ -30,11 +30,9 @@ I took your green background idea and changed it to an image.
 If we can get a decent font we might have something here.
 
 [Blake]
-Have been experiencing problems with my AVD Emulator. Went to Dr. Nicholsonja today and he
-helped me get it working. Not exactly sure what the issue was, but since its working now and
-I can test my code I am going to attempt to catch up to you on on workload.
-
-My plan is to follow the lead you have set. I feel like I have somewhat of a grip on how you have
-laid the activities out, and I am going to take a closer look at the different classes you have
-created. Ideally this weekend we could plan a time to discuss in more detail how we want the
-app to look, and function. I will email you some times when I will be free.
+The style activities are not complete, will continue working on them tomorrow night.
+Programatically, i dont think our cards array needs to have as many images as it does. Since
+suit does not matter, only number counts towards the score, we may be able to shorten the array
+by only having 1 of each card. Thats just an idea, it may or may not be useful especially as
+the png files are already loaded in the array. I will be occupied Sunday morning, but will
+continue working on the activities in the afternoon.
