+++ app/src/main/java/com/davidroach/blackjack_assignment1/GameActivity.java
@@ -2,20 +2,11 @@ package com.davidroach.blackjack_assignment1;
 
 import android.app.AlertDialog;
 import android.content.DialogInterface;
-import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.View;
 import android.view.Window;
 import android.widget.TextView;
 import android.widget.Button;
-import android.widget.ArrayAdapter;
-import android.widget.Spinner;
-import android.widget.LinearLayout;
-import android.widget.LinearLayout.LayoutParams;
-import android.widget.EditText;
-import android.content.res.Configuration;
-
 
 import org.w3c.dom.Text;
 
@@ -35,14 +26,13 @@ public class GameActivity extends AppCompatActivity {
     Player dealerObj;
     Player playerObj;
 
-    String winnerString;
+    String winnerString = "game result popup";
     String DEALER_WINS = "Dealer Wins.";
     String YOU_WIN = "You Win!!!";
-    String PUSH = "Push.";
 
     boolean dealerStands = false;
     boolean playerStands = false;
-
+    boolean playerWins;
     boolean dealerBusts;
     boolean playerBusts;
 
@@ -65,7 +55,7 @@ public class GameActivity extends AppCompatActivity {
 
 
         //init variables
-        betSize = 0;
+        betSize = 10;
         //playerHandScore = 0;
         //dealerHandScore = 0;
 
@@ -82,9 +72,6 @@ public class GameActivity extends AppCompatActivity {
         playerObj.chipCount = 100;
         /* Dealer does not need a chip count */
 
-
-
-
         //run main loop
         play();
         //gameResult();
@@ -93,68 +80,15 @@ public class GameActivity extends AppCompatActivity {
 
     public void betPopup(){
 
-
-
-        // 1. Instantiate an AlertDialog.Builder with its constructor
-        AlertDialog.Builder builder = new AlertDialog.Builder(GameActivity.this);
-
-        // 2. Chain together various setter methods to set the dialog characteristics
-        builder.setMessage(R.string.dialog_message)
-                .setTitle(R.string.dialog_title);
-
-        final EditText input = new EditText(GameActivity.this);
-        input.setText("1");
-
-        input.setRawInputType(Configuration.KEYBOARD_12KEY); //show number keybord for input
-
-        LinearLayout.LayoutParams lp = new LinearLayout.LayoutParams(
-                LinearLayout.LayoutParams.MATCH_PARENT,
-                LinearLayout.LayoutParams.MATCH_PARENT);
-        input.setLayoutParams(lp);
-        builder.setView(input);
-
-
-        builder.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int id) {
-
-                /* get value of edittext and assign it to betsize */
-                betSize = Integer.parseInt(input.getText().toString());
-                TextView betCountTv = (TextView)findViewById(R.id.bet_count_tv);
-                betCountTv.setText(Integer.toString(betSize));
-
-
-                setHandScore(playerObj);
-                setHandScore(dealerObj);
-            }
-        });
-
-
-
-
-        // 3. Get the AlertDialog from create()
-        AlertDialog dialog = builder.create();
-        dialog.show();
-
-
-
-
-
-
     }
 
 
 
     public void play(){
 
-        /*Place bet*/
-        betPopup();
-
         /* Bind to UI */
-
-
+        TextView betCountTv = (TextView)findViewById(R.id.bet_count_tv);
         TextView chipCountTv = (TextView)findViewById(R.id.chip_count_tv);
-        chipCountTv.setText(Integer.toString(playerObj.chipCount));
-
         TextView playerHandScore = (TextView)findViewById(R.id.player_score_tv);
         TextView dealerHandScore = (TextView)findViewById(R.id.dealer_score_tv);
         Button hitButton = (Button)findViewById(R.id.hit_button);
@@ -174,40 +108,46 @@ public class GameActivity extends AppCompatActivity {
 
        /*Initial deal*/
         deck.dealCard(playerObj);
+        setHandScore(playerObj);
         deck.dealCard(dealerObj);
+        setHandScore(dealerObj);
         deck.dealCard(playerObj);
+        setHandScore(playerObj);
         deck.dealCard(dealerObj);
+        setHandScore(dealerObj);
 
         /* check player hand for blackjack */
         if(hasBlackJack(playerObj) == true){
             winnerString = YOU_WIN;
-
+            playerWins = true;
 
         }
 
-        /* hit button onclick listener */
-        hitButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                takeHit(playerObj);
-            }
-        });
+        /* player takes hits here check playerStands before loop iteration */
 
-        /* standbutton onclick listener */
-        standButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                takeStand(playerObj);
-            }
-        });
+
+int i= 1+1;
+
+
+        /* Dealer takes hits until score is 16 or greater or he goes over 21 */
+
+
+        /* Compare Scores here */
+
+
+        /* Set winner string and show winner popup */
 
 
 
-        int i = 1+1; //debug stopper
 
 
+         gameResult();
 
-    }//end of play
+
+        /* Restart */
+        //restartGame();
+
+    }//end of plaqy
 
 
     public void setHandScore(Player playerIn){
@@ -229,15 +169,16 @@ public class GameActivity extends AppCompatActivity {
         }
 
 
+
+
+
+
     }
 
 
 
 
     public void gameResult(){
-
-        TextView chipCountTv = (TextView)findViewById(R.id.chip_count_tv);
-
         //show you win popup
         //remove bet size from player pot.
         AlertDialog.Builder builder = new AlertDialog.Builder(this);
@@ -251,21 +192,6 @@ public class GameActivity extends AppCompatActivity {
         AlertDialog alert = builder.create();
         alert.show();
 
-        if(winnerString.equals(YOU_WIN)){
-            /* Add betsize to chipcount, update chipcount text view*/
-            playerObj.chipCount += betSize;
-            chipCountTv.setText(Integer.toString(playerObj.chipCount));
-
-
-        }
-        else{
-            /*subtract betsize to chipcount, update chipcount text view */
-            playerObj.chipCount -= betSize;
-            chipCountTv.setText(Integer.toString(playerObj.chipCount));
-
-        }
-
-
 
         //restart game
         //restartGame();
@@ -276,37 +202,16 @@ public class GameActivity extends AppCompatActivity {
 
     public void takeHit(Player playerIn){
         deck.dealCard(playerIn);
-        setHandScore(playerIn);
-
-
-
-    }
-
-    public void dealerTurn(){
-        //dealer takes turns card and takes hits until >16 or bust
-        takeStand(dealerObj);
-    }
-
-    public void finishHand(){
-        compareScores(dealerObj,playerObj);
-
-        /*  check player wins boolean here */
-        gameResult();
 
 
     }
-
-
 
     public void takeStand(Player playerIn){
         if(Objects.equals("Dealer",playerIn.name)){
-            //dealerStands = true;
-            finishHand();
+            dealerStands = true;
         }
         else{
-            //playerStands = true;
-            dealerTurn();
-
+            playerStands = true;
         }
 
     }
@@ -333,21 +238,38 @@ public class GameActivity extends AppCompatActivity {
 
 
 
-    public void compareScores(Player dealerObj, Player playerObj){
-        if(dealerObj.handScore == playerObj.handScore){
-            winnerString = PUSH;
+
+    public void addToScore(Player playerIn, int cardValueIn){
+        //if card val is zero it is an ace.
+        //check and see if adding 11 would bust player
+        //if it will use the ace as a 1
+        if(cardValueIn ==0){
+
+            if((playerIn.handScore + cardValueIn) > 21){
+                playerIn.handScore ++;
+            }
+            else{
+                playerIn.handScore += 11;
+            }
 
         }
-        else if(dealerObj.handScore > playerObj.handScore || playerObj.handScore > 21){
-            //dealer wins
+        else{
+            playerIn.handScore += cardValueIn;
+        }
 
-            winnerString = DEALER_WINS;
+
+    }
+
+    public void compareScores(Player dealerObj, Player playerObj){
+
+        if(dealerObj.handScore > playerObj.handScore){
+            //dealer wins
+            playerWins = false;
 
         }
         else{
             //User wins
-
-            winnerString = YOU_WIN;
+            playerWins = true;
         }
