+++ app/src/main/res/layout/activity_game.xml
@@ -83,6 +83,7 @@
 
 
    <TextView
+       android:text="@string/your_score"
        android:layout_width="wrap_content"
        android:layout_height="wrap_content"
        android:id="@+id/player_score_tv"
