+++ app/src/main/res/layout/activity_start.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
     <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -8,44 +9,141 @@
         android:orientation="vertical"
         tools:context="com.davidroach.blackjack_assignment1.StartActivity"
         android:background="@drawable/background"
-        android:gravity="center_horizontal">
+        android:gravity="center_horizontal"
+    android:weightSum="1">
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="match_parent"
+        android:layout_height="174dp"
+        android:weightSum="3">
+
+        <ImageView
+            android:layout_width="69dp"
+            app:srcCompat="@drawable/lasvegas"
+            android:id="@+id/las_vegas_image"
+            android:layout_weight="2"
+            android:layout_height="188dp" />
 
         <TextView
             android:text="@string/blackjack"
+            android:textAlignment="center"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
+            android:layout_height="match_parent"
             android:id="@+id/blackjack_title_tv"
-            android:layout_centerHorizontal="true"
             android:textSize="36sp"
-            android:textColor="@android:color/white" />
+            android:textColor="@android:color/white"
+            android:layout_weight="1.41"
+            android:gravity="center_vertical" />
 
-    <Button
-        android:text="@string/play"
-        android:textSize="20sp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
-        android:background="@color/red"
-        android:id="@+id/play_button"
-        android:elevation="0dp" />
+    </LinearLayout>
 
-    <Button
-        android:text="@string/about"
-        android:textSize="20sp"
+    <LinearLayout
+        android:orientation="horizontal"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
-        android:background="@color/red"
-        android:id="@+id/about_button" />
+        android:layout_height="70dp"
+        android:weightSum="4">
 
-    <Button
-        android:text="@string/exit"
-        android:textSize="20sp"
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="70dp"
+            app:srcCompat="@drawable/club"
+            android:id="@+id/imageView7"
+            android:layout_weight="1"/>
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="70dp"
+            app:srcCompat="@drawable/diamond"
+            android:id="@+id/imageView6"
+            android:layout_weight="1"/>
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="70dp"
+            app:srcCompat="@drawable/spade"
+            android:id="@+id/imageView9"
+            android:layout_weight="1"/>
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="70dp"
+            app:srcCompat="@drawable/heart"
+            android:id="@+id/imageView8"
+            android:layout_weight="1"/>
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="vertical"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
-        android:background="@color/red"
-        android:id="@+id/exit_button" />
+        android:layout_height="216dp"
+        android:layout_weight="0.31">
+
+        <Button
+            android:text="@string/play"
+            android:textSize="20sp"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="20dp"
+            android:background="@color/red"
+            android:id="@+id/play_button"
+            android:elevation="0dp" />
+
+        <Button
+            android:text="@string/about"
+            android:textSize="20sp"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="20dp"
+            android:background="@color/red"
+            android:id="@+id/about_button" />
+
+        <Button
+            android:text="@string/exit"
+            android:textSize="20sp"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="20dp"
+            android:background="@color/red"
+            android:id="@+id/exit_button" />
+    </LinearLayout>
+
+    <LinearLayout
+        android:orientation="horizontal"
+        android:layout_width="match_parent"
+        android:layout_height="70dp"
+        android:weightSum="4">
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="70dp"
+            app:srcCompat="@drawable/spade"
+            android:id="@+id/imageView12"
+            android:layout_weight="1"/>
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="70dp"
+            app:srcCompat="@drawable/heart"
+            android:id="@+id/imageView13"
+            android:layout_weight="1"/>
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="70dp"
+            app:srcCompat="@drawable/club"
+            android:id="@+id/imageView10"
+            android:layout_weight="1"/>
+
+        <ImageView
+            android:layout_width="wrap_content"
+            android:layout_height="70dp"
+            app:srcCompat="@drawable/diamond"
+            android:id="@+id/imageView11"
+            android:layout_weight="1"/>
+
+    </LinearLayout>
 </LinearLayout>
