+++ app/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <resources>
     <string name="app_name">BlackJack_Assignment1</string>
     <string name="blackjack">BlackJack!</string>
-    <string name="objective">Beat the dealer by creating a hand that is closest to 21, but do not go
-        over. A player whos hand exceeds 21 automatically looses.</string>
+    <string name="objective">Try and make the sum of your cards higher than the sum of the
+        dealers cards, without going over 21.</string>
     <string name="david">David Roach</string>
     <string name="thomas">Thomas Crozier</string>
     <string name="dev_by">Developed By:</string>
@@ -18,7 +18,4 @@
     <string name="stand_label">Stand</string>
     <string name="dealer_score">Dealer:</string>
     <string name="your_score">You:</string>
-    <string name="dialog_title"></string>
-    <string name="dialog_message">Place Your bet:</string>
-    <string name="ok">Ok</string>
 </resources>
